requirements.txt
pandas==2.2.0
numpy==1.26.3
kagglehub==0.2.0
faker==22.5.1
dbt-snowflake==1.7.0
dbt-duckdb==1.7.0
scipy==1.11.4
tabulate==0.9.0
